+++ networking/libiproute/ipaddress.c
@@ -443,7 +443,7 @@ extern int ipaddr_list_or_flush(int argc, char **argv, int flush)
 	}
 
 	while (argc > 0) {
-		const unsigned short option_num = compare_string_array(option, *argv);
+		const int option_num = compare_string_array(option, *argv);
 		switch (option_num) {
 			case 0: /* to */
 				NEXT_ARG();
@@ -658,7 +658,7 @@ static int ipaddr_modify(int cmd, int argc, char **argv)
 	req.ifa.ifa_family = preferred_family;
 
 	while (argc > 0) {
-		const unsigned short option_num = compare_string_array(option, *argv);
+		const int option_num = compare_string_array(option, *argv);
 		switch (option_num) {
 			case 0: /* peer */
 			case 1: /* remote */
@@ -805,7 +805,7 @@ static int ipaddr_modify(int cmd, int argc, char **argv)
 extern int do_ipaddr(int argc, char **argv)
 {
 	const char *commands[] = { "add", "delete", "list", "show", "lst", "flush", 0 };
-	unsigned short command_num = 2;
+	int command_num = 2;
 
 	if (*argv) {
 		command_num = compare_string_array(commands, *argv);
